GDGT-2239 Consistent custom viz loading state
